+++ src/main/kotlin/com/github/saundefined/bitrix_idea/dialogs/CreateSimpleComponentDialogWrapper.kt
@@ -1,6 +1,7 @@
 package com.github.saundefined.bitrix_idea.dialogs
 
 import com.github.saundefined.bitrix_idea.BitrixIdeaBundle.message
+import com.github.saundefined.bitrix_idea.settings.AppSettingsState
 import com.github.saundefined.bitrix_idea.validation.ComponentSimpleCodeVerifier
 import com.intellij.ide.IdeView
 import com.intellij.ide.fileTemplates.FileTemplateManager
@@ -8,17 +9,28 @@ import com.intellij.ide.fileTemplates.FileTemplateUtil
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.DialogWrapper
 import com.intellij.psi.PsiDirectory
+import com.intellij.ui.ScrollPaneFactory
 import com.intellij.ui.dsl.builder.RowLayout
 import com.intellij.ui.dsl.builder.bindText
 import com.intellij.ui.dsl.builder.panel
 import com.intellij.ui.dsl.gridLayout.HorizontalAlign
+import com.intellij.ui.dsl.gridLayout.VerticalAlign
+import com.intellij.ui.table.JBTable
+import java.awt.BorderLayout
+import java.awt.Dimension
 import java.util.*
-import javax.swing.JComponent
+import javax.swing.*
+import javax.swing.table.DefaultTableModel
+import javax.swing.table.JTableHeader
 
 class CreateSimpleComponentDialogWrapper : DialogWrapper(true) {
+    val settings: AppSettingsState
+        get() = AppSettingsState().getInstance()
+    
     var code: String = ""
     var name: String = ""
     var description: String = ""
+    var languages: List<String> = settings.languages
 
     lateinit var view: IdeView
     lateinit var project: Project
@@ -30,6 +42,74 @@ class CreateSimpleComponentDialogWrapper : DialogWrapper(true) {
     }
 
     override fun createCenterPanel(): JComponent {
+        val tableModel = object : DefaultTableModel(
+            languages.map { arrayOf(it) }.toTypedArray(),
+            arrayOf(message("localization.table.column.name"))
+        ) {
+            override fun isCellEditable(row: Int, column: Int): Boolean {
+                return false
+            }
+        }
+
+        val table = JBTable(tableModel)
+            .apply {
+                tableHeader = JTableHeader(this.columnModel).apply {
+                    resizingAllowed = false
+                    reorderingAllowed = false
+                }
+            }
+
+        val buttonDimension = Dimension(24, 24)
+
+        val addButton = JButton("+").apply {
+            preferredSize = buttonDimension
+            minimumSize = buttonDimension
+            maximumSize = buttonDimension
+
+            addActionListener {
+                val result =
+                    JOptionPane.showInputDialog(
+                        null,
+                        message("localization.add"),
+                        message("localization.add.title"),
+                        JOptionPane.QUESTION_MESSAGE,
+                    )
+
+                if (!result.isNullOrEmpty()) {
+                    tableModel.addRow(arrayOf(result))
+                    languages = tableModel.dataVector.map { (it as Vector<*>)[0].toString() }
+                }
+
+                this@CreateSimpleComponentDialogWrapper.pack()
+            }
+        }
+
+        val removeButton = JButton("-").apply {
+            preferredSize = buttonDimension
+            minimumSize = buttonDimension
+            maximumSize = buttonDimension
+
+            addActionListener {
+                if (table.selectedRow != -1 && tableModel.rowCount > 1) {
+                    tableModel.removeRow(table.selectedRow)
+                    languages = tableModel.dataVector.map { (it as Vector<*>)[0].toString() }
+                }
+
+                this@CreateSimpleComponentDialogWrapper.pack()
+            }
+        }
+
+        val btnPanel = JPanel().apply {
+            layout = BoxLayout(this, BoxLayout.Y_AXIS)
+
+            add(addButton)
+            add(removeButton)
+        }
+
+        val wrapperPanel = JPanel(BorderLayout())
+        wrapperPanel.add(btnPanel, BorderLayout.LINE_END)
+        wrapperPanel.add(ScrollPaneFactory.createScrollPane(table), BorderLayout.CENTER)
+
         return panel {
             row {
                 label(message("create.component.simple.code"))
@@ -73,6 +153,14 @@ class CreateSimpleComponentDialogWrapper : DialogWrapper(true) {
                         }
                     }
             }.layout(RowLayout.PARENT_GRID)
+
+            row {
+                label(message("localization.table.row.name"))
+                    .verticalAlign(VerticalAlign.TOP)
+                cell(wrapperPanel)
+                    .horizontalAlign(HorizontalAlign.FILL)
+            }
+                .layout(RowLayout.PARENT_GRID)
         }
     }
 
@@ -117,8 +205,6 @@ class CreateSimpleComponentDialogWrapper : DialogWrapper(true) {
                 properties,
                 defaultTemplateDirectory
             )
-
-            val languages = listOf("ru", "en")
 
             val langDirectory = componentDirectory.createSubdirectory("lang")
             languages.forEach { language ->